backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.1
Flask-JWT-Extended==4.6.0
google-auth==2.22.0
